test_project/pyproject.toml
[project]
name = "testnats"
version = "0.1.0"
dynamic = ["dependencies", "classifiers"]
requires-python = "~=3.12"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
classifiers = [
  "Private :: Do not Upload",
]

[tool.poetry.requires-plugins]
poetry-plugin-dotenv = "2.8.0"

[tool.poetry.plugins.dotenv]
ignore = "false"
location = "config/.env"

[tool.poetry.dependencies]
django = "5.2.3"
django-redis = "6.0.0"
nats-py = "2.12.0"
uvicorn = "0.38.0"
uvicorn-worker = "0.4.0"
pyjwt = "^2.10.1"
cryptography = "^46.0.3"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^1.4"
ruff = "^0.12"

[tool.ruff]
# Ruff config: https://docs.astral.sh/ruff/settings
preview = true
fix = true
target-version = "py312"
line-length = 80
exclude = ["migrations"]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = false

[tool.ruff.lint]
select = [
  "A",    # flake8-builtins
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "C90",  # maccabe
  "COM",  # flake8-commas
  "D",    # pydocstyle
  "DTZ",  # flake8-datetimez
  "E",    # pycodestyle
  "ERA",  # flake8-eradicate
  "EXE",  # flake8-executable
  "F",    # pyflakes
  "FA",   # flake8-future-annotations
  "FBT",  # flake8-boolean-trap
  "FLY",  # pyflint
  "FURB", # refurb
  "G",    # flake8-logging-format
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PL",   # pylint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "Q",    # flake8-quotes
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # ruff
  "S",    # flake8-bandit
  "SIM",  # flake8-simpify
  "SLF",  # flake8-self
  "SLOT", # flake8-slots
  "T100", # flake8-debugger
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
ignore = [
  "A005",   # allow to shadow stdlib and builtin module names
  "COM812", # trailing comma, conflicts with `ruff format`
  # Different doc rules that we don't really care about:
  "D100",
  "D104",
  "D106",
  "D203",
  "D212",
  "D401",
  "D404",
  "D405",
  "ISC001",  # implicit string concat conflicts with `ruff format`
  "ISC003",  # prefer explicit string concat over implicit concat
  "PLR09",   # we have our own complexity rules
  "PLR2004", # do not report magic numbers
  "PLR6301", # do not require classmethod / staticmethod when self not used
  "TRY003",  # long exception messages from `tryceratops`
  # Octolo
  "D205",
  "D400",
  "E501",
  "D102",
  "S101",
  # "E800", this one bug ?
]
external = [ "WPS" ]

# Plugin configs:
flake8-import-conventions.banned-from = [ "ast" ]
flake8-quotes.inline-quotes = "single"
mccabe.max-complexity = 6
pydocstyle.convention = "google"

# TODO: Fix as it is not applied
[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["D101", "E501", "RUF012"]
"tests/*.py" = ["S101"]
